Fix false negative test for new Explore plot tabs for annotation scatter
