+++ app/javascript/components/explore/PlotTabs.jsx
@@ -1,5 +1,20 @@
 import React from 'react'
 
+const tabList = [
+  { key: 'loading', label: 'Loading...' },
+  { key: 'scatter', label: 'Scatter' },
+  { key: 'annotatedScatter', label: 'Annotated scatter' },
+  { key: 'correlatedScatter', label: 'Correlation' },
+  { key: 'distribution', label: 'Distribution' },
+  { key: 'dotplot', label: 'Dot plot' },
+  { key: 'heatmap', label: 'Heatmap' },
+  { key: 'geneListHeatmap', label: 'Precomputed heatmap' },
+  { key: 'spatial', label: 'Spatial' },
+  { key: 'genome', label: 'Genome' },
+  { key: 'infercnv-genome', label: 'Genome (inferCNV)' },
+  { key: 'images', label: 'Images' }
+]
+
 const disabledTooltips = {
   'annotatedScatter': { numToSearch: '1', isMulti: false },
   'scatter': { numToSearch: '1', isMulti: false },
@@ -14,7 +29,7 @@ const disabledTooltips = {
  * Responsible for shows tabs available for a given view of the study
 */
 export default function PlotTabs({
-  shownTab, enabledTabs, disabledTabs, tabList, updateExploreParams,
+  shownTab, enabledTabs, disabledTabs, updateExploreParams,
   isNewExploreUX
 }) {
   return (
